Added sound volume triggers and changed input methods
Added in the volume triggers for player movement, smallest is sneak, largest is running and when the player is still there is no volume trigger.

Changed input methods so that it can track if the user is still/running/walking/sneaking using booleans and added a script that checks for player interaction with stairs (the script has to be attached to the stairs object/prefab to work).

Generally fixed some errors with the player script that had gone unnoticed.